refactor: seperate dummy data
