+++ data/dummyData.js
@@ -1,0 +1,25 @@
+const dummyData = [
+  {
+    title: "iphone"
+  },
+  {
+    title: "ipad"
+  },
+  {
+    title: "imac"
+  },
+  {
+    title: "ipod"
+  },
+  {
+    title: "iphoneX"
+  },
+  {
+    title: "iphone6"
+  },
+  {
+    title: "apple watch"
+  }
+];
+
+export default dummyData;
